requirements.txt
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
MouseInfo==0.1.3
pillow==10.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python3-xlib==0.15
pytweening==1.2.0
pywhatkit==5.4
requests==2.31.0
soupsieve==2.5
urllib3==2.2.1
Werkzeug==3.0.2
wikipedia==1.4.0
zipp==3.18.1
